// src/app/features/payments/model/payment-list-params.ts
import { EntityListParams } from '../../../shared/models/common/entity-list-params';

type PaymentPurpose =
  | 0 //CreditPackage
  | 1; //SchoolSubscription

type PaymentMethod =
  | 0 //VNPAY
  | 1; //PayOS
type PaymentStatus =
  | 0 //Pending
  | 1; //Paid

export interface PaymentListParams extends EntityListParams {
  paymentPurpose?: PaymentPurpose;
  paymentMethod?: PaymentMethod;
  paymentStatus?: PaymentStatus;
}